// sdks/javascript/src/models/branding-preference.ts
export interface BrandingLayout {
  activeLayout?: string;
  sideImg?: {
    altText?: string;
    imgURL?: string;
  };
}

/**
 * Interface for organization details configuration.
 */
export interface BrandingOrganizationDetails {
  displayName?: string;
  supportEmail?: string;
}

/**
 * Interface for URL configurations.
 */
export interface UrlsConfig {
  cookiePolicyURL?: string;
  privacyPolicyURL?: string;
  selfSignUpURL?: string;
  termsOfUseURL?: string;
}

/**
 * Interface for branding preference theme configuration.
 */
export interface BrandingTheme {
  DARK?: ThemeVariant;
  LIGHT?: ThemeVariant;
  activeTheme?: string;
}

/**
 * Interface for branding preference configuration.
 */
export interface BrandingPreferenceConfig {
  configs?: {
    isBrandingEnabled?: boolean;
    removeDefaultBranding?: boolean;
    selfSignUpEnabled?: boolean;
  };
  layout?: BrandingLayout;
  organizationDetails?: BrandingOrganizationDetails;
  theme?: BrandingTheme;
  urls?: UrlsConfig;
}

/**
 * Interface for branding preference configuration.
 */
export interface BrandingPreference {
  locale?: string;
  name?: string;
  preference?: BrandingPreferenceConfig;
  type?: string;
}